+++ src/lib.rs
@@ -39,7 +39,7 @@ pub use collections::*;
 pub use value::*;
 pub use convert::*;
 use crate::lang::{Sequence};
-use crate::compiler::{RantCompiler, Reporter};
+use crate::compiler::{RantCompiler, Reporter, ErrorKind as CompilerErrorKind};
 use crate::runtime::*;
 use std::{path::Path, rc::Rc, cell::RefCell};
 use random::RantRng;
@@ -113,25 +113,25 @@ impl Default for Rant {
 impl Rant {
   /// Compiles a source string using the specified reporter.
   #[must_use = "compiling a program without storing or running it achieves nothing"]
-  pub fn compile<R: Reporter>(&self, source: &str, reporter: &mut R) -> Result<RantProgram, ()> {
+  pub fn compile<R: Reporter>(&self, source: &str, reporter: &mut R) -> Result<RantProgram, CompilerErrorKind> {
     RantCompiler::compile_string(source, reporter)
   }
 
   /// Compiles a source string without reporting problems.
   #[must_use = "compiling a program without storing or running it achieves nothing"]
-  pub fn compile_quiet(&self, source: &str) -> Result<RantProgram, ()> {
+  pub fn compile_quiet(&self, source: &str) -> Result<RantProgram, CompilerErrorKind> {
     RantCompiler::compile_string(source, &mut ())
   }
   
   /// Compiles a source file using the specified reporter.
   #[must_use = "compiling a program without storing or running it achieves nothing"]
-  pub fn compile_file<P: AsRef<Path>, R: Reporter>(&self, path: P, reporter: &mut R) -> Result<RantProgram, ()> {
+  pub fn compile_file<P: AsRef<Path>, R: Reporter>(&self, path: P, reporter: &mut R) -> Result<RantProgram, CompilerErrorKind> {
     RantCompiler::compile_file(path, reporter)
   }
 
   /// Compiles a source file without reporting problems.
   #[must_use = "compiling a program without storing or running it achieves nothing"]
-  pub fn compile_file_quiet<P: AsRef<Path>>(&self, path: P) -> Result<RantProgram, ()> {
+  pub fn compile_file_quiet<P: AsRef<Path>>(&self, path: P) -> Result<RantProgram, CompilerErrorKind> {
     RantCompiler::compile_file(path, &mut ())
   }