+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added new `Rant::with_options` constructor that lets you customize the context using new `RantOptions` struct.
 * Added `RantValueType` to represent the type of a value.
+* Replaced empty `Err` variant on `Rant::compile*` methods with new `compiler::ErrorKind` enum.
 * Changed `IndexError` and `KeyError` to use `RantValueType` instead of `&'static str` to represent types.
 * Renamed `get_by_index`, `set_by_index`, `get_by_key`, and `set_by_key` methods in `RantValue` to `index_get`, `index_set`, `key_get`, and `key_set` respectively.
 * Expanded type aliases on `Rant::compile*` method return types to make them less confusing.